started creating Parent classes.
Created a parent class for platforms which abstracts common functions.
This isn't finished yet but will simplify the platform classes.

Also created a task class so that tasks can be well defined objects
with the same properties and methods.
Task properties that get uploaded to platform APIs will be common
but converted to use the platform's notation before making requests.